refactor(test): Restructure test app
