--- test/apps/_flask/app.py
@@ -1,3 +1,0 @@
-from . import create_openapi_app
-
-app = create_openapi_app()
