+++ test/apps/openapi/__init__.py
@@ -1,0 +1,1 @@
+from . import _aiohttp, _fastapi, _flask, schema
