+++ test/apps/openapi/_fastapi/app.py
@@ -1,0 +1,3 @@
+from . import create_app
+
+app = create_app()
