+++ test/apps/openapi/_flask/app.py
@@ -1,0 +1,3 @@
+from . import create_app
+
+app = create_app()
